backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pandas
numpy
python-dateutil
scikit-learn  
nltk   